+++ ost/__init__.py
@@ -1,4 +1,3 @@
-from ost.s1.s1scene import Sentinel1Scene
-from ost.Project import Generic, Sentinel1, Sentinel1Batch
+from ost.Project import Sentinel1Batch
 
-__all__ = ['Sentinel1Scene', 'Sentinel1Batch', 'Sentinel1', 'Generic']
+__all__ = ['Sentinel1Batch']